+++ cli/cli/commands/service/logs/logs.go
@@ -107,7 +107,7 @@ func run(
 
 	clusterConfig, err := kurtosis_config_getter.GetKurtosisClusterConfig()
 	if err != nil {
-		return stacktrace.Propagate(err, "And error occurred getting the Kurtosis cluster config")
+		return stacktrace.Propagate(err, "An error occurred getting the Kurtosis cluster config")
 	}
 
 	clusterType := clusterConfig.GetClusterType()
@@ -165,7 +165,8 @@ func run(
 		return stacktrace.Propagate(err, "An error occurred connecting to the local Kurtosis engine")
 	}
 
-	userServiceLogsByGuidChan, cancelStreamUserServiceLogsFunc, err := kurtosisCtx.GetUserServiceLogs(ctx, enclaveId, userServiceGuids, shouldFollowLogs)
+	//TODO handle notFoundUserServiceGuidsChan case printing a message to the user
+	userServiceLogsByGuidChan, _, cancelStreamUserServiceLogsFunc, err := kurtosisCtx.GetUserServiceLogs(ctx, enclaveId, userServiceGuids, shouldFollowLogs)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred getting user service logs from user services with GUIDs '%+v' in enclave '%v' and with follow logs value '%v'", userServiceGuids, enclaveId, shouldFollowLogs)
 	}