+++ engine/server/engine/centralized_logs/loki_logs_database_client_test.go
@@ -6,6 +6,7 @@ import (
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/enclave"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/service"
 	"github.com/kurtosis-tech/kurtosis/engine/server/engine/centralized_logs/mocks"
+	"github.com/kurtosis-tech/stacktrace"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 	"io"
@@ -13,6 +14,7 @@ import (
 	"regexp"
 	"strings"
 	"testing"
+	"time"
 )
 
 const (
@@ -38,6 +40,11 @@ const (
 	expectedEntriesLimitQueryParamValue     = "4000"
 	expectedDirectionQueryParamValue        = "forward"
 	expectedAmountQueryParams               = 4
+
+	userServiceContainerType = "user-service"
+
+	testTimeOut = 30 * time.Second
+
 )
 
 func TestGetUserServiceLogs_ValidResponse(t *testing.T) {
@@ -114,29 +121,55 @@ func TestGetUserServiceLogs_ValidResponse(t *testing.T) {
 
 	ctx := context.Background()
 
-	userServiceLogsByUserServiceGuids, err := logsDatabaseClient.GetUserServiceLogs(ctx, enclaveId, userServiceGuids)
-	require.NoError(t, err, "An error occurred getting user service logs for GUIDs '%+v' in enclave '%v'", userServiceGuids, enclaveId)
-	require.NotNil(t, userServiceLogsByUserServiceGuids, "Received a nil user service logs, but a non-nil value was expected")
-
-	require.Equal(t, len(userServiceGuids), len(userServiceLogsByUserServiceGuids))
-
 	expectedUserServiceAmountLogLinesByUserServiceGuid := map[service.ServiceGUID]int{
 		testUserService1Guid: 3,
 		testUserService2Guid: 4,
 		testUserService3Guid: 2,
 	}
 
-	for userServiceGuid := range userServiceGuids {
-		logLines, found := userServiceLogsByUserServiceGuids[userServiceGuid]
-		require.True(t, found)
+	userServiceLogsByGuidChan, errChan, closeStreamFunc, err := logsDatabaseClient.GetUserServiceLogs(ctx, enclaveId, userServiceGuids)
+	defer closeStreamFunc()
 
-		expectedAmountLogLines, found := expectedUserServiceAmountLogLinesByUserServiceGuid[userServiceGuid]
-		require.True(t, found)
+	require.NoError(t, err, "An error occurred getting user service logs for GUIDs '%+v' in enclave '%v'", userServiceGuids, enclaveId)
+	require.NotNil(t, userServiceLogsByGuidChan, "Received a nil user service logs channel, but a non-nil value was expected")
+	require.Nil(t, errChan, "Received a not nil error channel, but a nil value was expected")
 
-		require.Equal(t, expectedAmountLogLines, len(logLines))
+	var testEvaluationErr error
 
-		require.Equal(t, expectedFirstLogLineOnEachService, logLines[0].GetContent())
+	shouldReceiveStream := true
+	for shouldReceiveStream {
+		select {
+		case <-time.Tick(testTimeOut):
+			testEvaluationErr = stacktrace.NewError("Receiving stream logs in the test has reached the '%v' time out", testTimeOut)
+			shouldReceiveStream = false
+			break
+		case userServiceLogsByGuid, isChanOpen := <-userServiceLogsByGuidChan:
+			if !isChanOpen {
+				shouldReceiveStream = false
+				break
+			}
+
+			require.Equal(t, len(userServiceGuids), len(userServiceLogsByGuid))
+
+			for userServiceGuid := range userServiceGuids {
+				logLines, found := userServiceLogsByGuid[userServiceGuid]
+				require.True(t, found)
+
+				expectedAmountLogLines, found := expectedUserServiceAmountLogLinesByUserServiceGuid[userServiceGuid]
+				require.True(t, found)
+
+				require.Equal(t, expectedAmountLogLines, len(logLines))
+
+				require.Equal(t, expectedFirstLogLineOnEachService, logLines[0].GetContent())
+			}
+
+			shouldReceiveStream = false
+			break
+		}
 	}
+
+	require.NoError(t, testEvaluationErr)
+
 }
 
 func TestNewUserServiceLogLinesByUserServiceGuidFromLokiStreamsReturnSuccessfullyForLogTailJsonResponseBody(t *testing.T) {
@@ -145,55 +178,32 @@ func TestNewUserServiceLogLinesByUserServiceGuidFromLokiStreamsReturnSuccessfull
 	userServiceGuidStr := "stream-logs-test-service-1666785469"
 	userServiceGuid := service.ServiceGUID(userServiceGuidStr)
 
+	expectedValuesInStream1 := [][]string{
+		{"1666785473000000000", "{\"container_id\":\"b0735bc50a76a0476928607aca13a4c73c814036bdbf8b989c2f3b458cc21eab\",\"container_name\":\"/ts-testsuite.stream-logs-test.1666785464--user-service--stream-logs-test-service-1666785469\",\"source\":\"stdout\",\"log\":\"kurtosis\",\"comKurtosistechGuid\":\"stream-logs-test-service-1666785469\",\"comKurtosistechContainerType\":\"user-service\",\"com.kurtosistech.enclave-id\":\"ts-testsuite.stream-logs-test.1666785464\"}"},
+	}
+
+	expectedValuesInStream2 := [][]string{
+		{"1666785473000000000", "{\"comKurtosistechGuid\":\"stream-logs-test-service-1666785469\",\"container_id\":\"b0735bc50a76a0476928607aca13a4c73c814036bdbf8b989c2f3b458cc21eab\",\"container_name\":\"/ts-testsuite.stream-logs-test.1666785464--user-service--stream-logs-test-service-1666785469\",\"source\":\"stdout\",\"log\":\"test\",\"comKurtosistechContainerType\":\"user-service\",\"com.kurtosistech.enclave-id\":\"ts-testsuite.stream-logs-test.1666785464\"}"},
+	}
+
+	expectedValuesInStream3 := [][]string{
+		{"1666785473000000000", "{\"comKurtosistechContainerType\":\"user-service\",\"com.kurtosistech.enclave-id\":\"ts-testsuite.stream-logs-test.1666785464\",\"comKurtosistechGuid\":\"stream-logs-test-service-1666785469\",\"container_id\":\"b0735bc50a76a0476928607aca13a4c73c814036bdbf8b989c2f3b458cc21eab\",\"container_name\":\"/ts-testsuite.stream-logs-test.1666785464--user-service--stream-logs-test-service-1666785469\",\"source\":\"stdout\",\"log\":\"running\"}"},
+	}
+
+	expectedValuesInStream4 := [][]string{
+		{"1666785473000000000", "{\"container_name\":\"/ts-testsuite.stream-logs-test.1666785464--user-service--stream-logs-test-service-1666785469\",\"source\":\"stdout\",\"log\":\"successfully\",\"comKurtosistechGuid\":\"stream-logs-test-service-1666785469\",\"comKurtosistechContainerType\":\"user-service\",\"com.kurtosistech.enclave-id\":\"ts-testsuite.stream-logs-test.1666785464\",\"container_id\":\"b0735bc50a76a0476928607aca13a4c73c814036bdbf8b989c2f3b458cc21eab\"}"},
+	}
+
+	lokiStreams1 := newLokiStreamValueForTest(userServiceGuid, expectedValuesInStream1)
+	lokiStreams2 := newLokiStreamValueForTest(userServiceGuid, expectedValuesInStream2)
+	lokiStreams3 := newLokiStreamValueForTest(userServiceGuid, expectedValuesInStream3)
+	lokiStreams4 := newLokiStreamValueForTest(userServiceGuid, expectedValuesInStream4)
+
 	lokiStreams := []lokiStreamValue{
-		{
-			Stream: struct {
-				KurtosisContainerType string `json:"comKurtosistechContainerType"`
-				KurtosisGUID          string `json:"comKurtosistechGuid"`
-			}(struct {
-				KurtosisContainerType string
-				KurtosisGUID          string
-			}{KurtosisContainerType: "user-service", KurtosisGUID: userServiceGuidStr}),
-			Values: [][]string{
-				{"1666785473000000000", "{\"container_id\":\"b0735bc50a76a0476928607aca13a4c73c814036bdbf8b989c2f3b458cc21eab\",\"container_name\":\"/ts-testsuite.stream-logs-test.1666785464--user-service--stream-logs-test-service-1666785469\",\"source\":\"stdout\",\"log\":\"kurtosis\",\"comKurtosistechGuid\":\"stream-logs-test-service-1666785469\",\"comKurtosistechContainerType\":\"user-service\",\"com.kurtosistech.enclave-id\":\"ts-testsuite.stream-logs-test.1666785464\"}"},
-			},
-		},
-		{
-			Stream: struct {
-				KurtosisContainerType string `json:"comKurtosistechContainerType"`
-				KurtosisGUID          string `json:"comKurtosistechGuid"`
-			}(struct {
-				KurtosisContainerType string
-				KurtosisGUID          string
-			}{KurtosisContainerType: "user-service", KurtosisGUID: userServiceGuidStr}),
-			Values: [][]string{
-				{"1666785473000000000", "{\"comKurtosistechGuid\":\"stream-logs-test-service-1666785469\",\"container_id\":\"b0735bc50a76a0476928607aca13a4c73c814036bdbf8b989c2f3b458cc21eab\",\"container_name\":\"/ts-testsuite.stream-logs-test.1666785464--user-service--stream-logs-test-service-1666785469\",\"source\":\"stdout\",\"log\":\"test\",\"comKurtosistechContainerType\":\"user-service\",\"com.kurtosistech.enclave-id\":\"ts-testsuite.stream-logs-test.1666785464\"}"},
-			},
-		},
-		{
-			Stream: struct {
-				KurtosisContainerType string `json:"comKurtosistechContainerType"`
-				KurtosisGUID          string `json:"comKurtosistechGuid"`
-			}(struct {
-				KurtosisContainerType string
-				KurtosisGUID          string
-			}{KurtosisContainerType: "user-service", KurtosisGUID: userServiceGuidStr}),
-			Values: [][]string{
-				{"1666785473000000000", "{\"comKurtosistechContainerType\":\"user-service\",\"com.kurtosistech.enclave-id\":\"ts-testsuite.stream-logs-test.1666785464\",\"comKurtosistechGuid\":\"stream-logs-test-service-1666785469\",\"container_id\":\"b0735bc50a76a0476928607aca13a4c73c814036bdbf8b989c2f3b458cc21eab\",\"container_name\":\"/ts-testsuite.stream-logs-test.1666785464--user-service--stream-logs-test-service-1666785469\",\"source\":\"stdout\",\"log\":\"running\"}"},
-			},
-		},
-		{
-			Stream: struct {
-				KurtosisContainerType string `json:"comKurtosistechContainerType"`
-				KurtosisGUID          string `json:"comKurtosistechGuid"`
-			}(struct {
-				KurtosisContainerType string
-				KurtosisGUID          string
-			}{KurtosisContainerType: "user-service", KurtosisGUID: userServiceGuidStr}),
-			Values: [][]string{
-				{"1666785473000000000", "{\"container_name\":\"/ts-testsuite.stream-logs-test.1666785464--user-service--stream-logs-test-service-1666785469\",\"source\":\"stdout\",\"log\":\"successfully\",\"comKurtosistechGuid\":\"stream-logs-test-service-1666785469\",\"comKurtosistechContainerType\":\"user-service\",\"com.kurtosistech.enclave-id\":\"ts-testsuite.stream-logs-test.1666785464\",\"container_id\":\"b0735bc50a76a0476928607aca13a4c73c814036bdbf8b989c2f3b458cc21eab\"}"},
-			},
-		},
+		lokiStreams1,
+		lokiStreams2,
+		lokiStreams3,
+		lokiStreams4,
 	}
 
 	resultLogsByKurtosisUserServiceGuid, err := newUserServiceLogLinesByUserServiceGuidFromLokiStreams(lokiStreams)
@@ -296,4 +306,24 @@ func TestFilterExistingServiceGuids_UnexpectedResponseObjectShape(t *testing.T) 
 	result, err := lokiDbClient.FilterExistingServiceGuids(ctx, testEnclaveId, requestedServiceGuids)
 	require.Nil(t, result)
 	require.Contains(t, err.Error(), "The logs database returns an error status when fetching the existing service GUIDs. Response was: ")
+}
+
+
+// ====================================================================================================
+//
+//	Private Helper Functions
+//
+// ====================================================================================================
+func newLokiStreamValueForTest(userServiceGuid service.ServiceGUID, expectedValues [][]string) lokiStreamValue {
+	newLokiStreamValue := lokiStreamValue{
+		Stream: struct {
+			KurtosisContainerType string `json:"comKurtosistechContainerType"`
+			KurtosisGUID          string `json:"comKurtosistechGuid"`
+		}(struct {
+			KurtosisContainerType string
+			KurtosisGUID          string
+		}{KurtosisContainerType: userServiceContainerType, KurtosisGUID: string(userServiceGuid)}),
+		Values: expectedValues,
+	}
+	return newLokiStreamValue
 }