+++ internal_testsuites/golang/testsuite/stream_logs_test/stream_logs_test.go
@@ -8,6 +8,7 @@ package stream_logs_test
 import (
 	"context"
 	"github.com/kurtosis-tech/kurtosis-cli/golang_internal_testsuite/test_helpers"
+	"github.com/kurtosis-tech/kurtosis/api/golang/core/lib/enclaves"
 	"github.com/kurtosis-tech/kurtosis/api/golang/core/lib/services"
 	"github.com/kurtosis-tech/kurtosis/api/golang/engine/lib/kurtosis_context"
 	"github.com/kurtosis-tech/stacktrace"
@@ -23,31 +24,29 @@ const (
 	dockerGettingStartedImage                    = "docker/getting-started"
 	exampleServiceId          services.ServiceID = "stream-logs"
 
-	exampleServicePortId         = "http"
-	exampleServicePrivatePortNum = 80
 
 	waitForAllLogsBeingCollectedInSeconds = 2
 
-	testTimeOut = 2 * time.Minute
+	testTimeOut = 90 * time.Second
 
-	shouldFollowLogsFirstRequest  = false
-	shouldFollowLogsSecondRequest = true
+	shouldFollowLogs  = true
+	shouldNotFollowLogs = false
+
+	nonExistentUserServiceGuid = "stream-logs-1667939326"
 )
 
-var (
-	shouldFollowLogsRequestOptions = []bool{shouldFollowLogsFirstRequest, shouldFollowLogsSecondRequest}
-
-	exampleServicePrivatePortSpec = services.NewPortSpec(
-		exampleServicePrivatePortNum,
-		services.PortProtocol_TCP,
-	)
-)
+type getUserServiceLogsRequestInfoAndExpectedResults struct {
+	requestedEnclaveID enclaves.EnclaveID
+	requestedUserServiceGuids map[services.ServiceGUID]bool
+	requestedFollowLogs bool
+	expectedLogLines []string
+	expectedNotFoundServiceGuids map[services.ServiceGUID]bool
+}
 
 func TestStreamLogs(t *testing.T) {
 	ctx := context.Background()
 
-	expectedLogLines := []string{"kurtosis", "test", "running", "successfully"}
-	expectedAmountOfLogLines := 4
+	expectedLogLineValues := []string{"kurtosis", "test", "running", "successfully"}
 
 	// ------------------------------------- ENGINE SETUP ----------------------------------------------
 	enclaveCtx, stopEnclaveFunc, _, err := test_helpers.CreateEnclave(t, ctx, testName, isPartitioningEnabled)
@@ -66,42 +65,85 @@ func TestStreamLogs(t *testing.T) {
 
 	enclaveID := enclaveCtx.GetEnclaveID()
 
-	time.Sleep(waitForAllLogsBeingCollectedInSeconds * time.Second)
-
 	serviceCtx, err := enclaveCtx.GetServiceContext(exampleServiceId)
 	require.NoError(t, err)
 
 	userServiceGuid := serviceCtx.GetServiceGUID()
 
-	userServiceGUIDs := map[services.ServiceGUID]bool{
+	userServiceGuids := map[services.ServiceGUID]bool{
 		userServiceGuid: true,
+	}
+
+	emptyUserServiceGuids := map[services.ServiceGUID]bool{}
+
+	nonExistentUserServiceGuids := map[services.ServiceGUID]bool{
+		nonExistentUserServiceGuid: true,
 	}
 
 	time.Sleep(waitForAllLogsBeingCollectedInSeconds * time.Second)
 
-	// will test executing getUserServiceLogs with shouldFollowLogs = false in the first iteration,
-	// and with shouldFollowLogs = true (which is used to tail logs) in the second iteration
-	for _, shouldFollowLogs := range shouldFollowLogsRequestOptions {
-		//TODO handle notFOundChannel
-		userServiceLogsByGuidChan, _, cancelStreamUserServiceLogsFunc, err := kurtosisCtx.GetUserServiceLogs(ctx, enclaveID, userServiceGUIDs, shouldFollowLogs)
+	firstCallRequestInfoAndExpectedResults := &getUserServiceLogsRequestInfoAndExpectedResults{
+		requestedEnclaveID:           enclaveID,
+		requestedUserServiceGuids:    userServiceGuids,
+		requestedFollowLogs:          shouldNotFollowLogs,
+		expectedLogLines:             expectedLogLineValues,
+		expectedNotFoundServiceGuids: emptyUserServiceGuids,
+
+	}
+
+	secondCallRequestInfoAndExpectedResults := &getUserServiceLogsRequestInfoAndExpectedResults{
+		requestedEnclaveID:           enclaveID,
+		requestedUserServiceGuids:    userServiceGuids,
+		requestedFollowLogs:          shouldFollowLogs,
+		expectedLogLines:             expectedLogLineValues,
+		expectedNotFoundServiceGuids: emptyUserServiceGuids,
+	}
+
+	thirdCallRequestInfoAndExpectedResults := &getUserServiceLogsRequestInfoAndExpectedResults{
+		requestedEnclaveID: enclaveID,
+		requestedUserServiceGuids: nonExistentUserServiceGuids,
+		requestedFollowLogs: shouldFollowLogs,
+		expectedNotFoundServiceGuids: nonExistentUserServiceGuids,
+	}
+
+	getUserServiceLogsRequestInfoAndExpectedResultsList := []*getUserServiceLogsRequestInfoAndExpectedResults{
+		firstCallRequestInfoAndExpectedResults,
+		secondCallRequestInfoAndExpectedResults,
+		thirdCallRequestInfoAndExpectedResults,
+	}
+
+	for _, userServiceLogsRequestInfoAndExpectedResults := range getUserServiceLogsRequestInfoAndExpectedResultsList {
+
+		requestedEnclaveId := userServiceLogsRequestInfoAndExpectedResults.requestedEnclaveID
+		requestedUserServiceGuids := userServiceLogsRequestInfoAndExpectedResults.requestedUserServiceGuids
+		requestedShouldFollowLogs := userServiceLogsRequestInfoAndExpectedResults.requestedFollowLogs
+		expectedLogLines := userServiceLogsRequestInfoAndExpectedResults.expectedLogLines
+		expectedNonExistenceUserServiceGuids := userServiceLogsRequestInfoAndExpectedResults.expectedNotFoundServiceGuids
+
+		userServiceLogsByGuidChan, notFoundUserServiceGuidsChan, cancelStreamUserServiceLogsFunc, err := kurtosisCtx.GetUserServiceLogs(ctx, requestedEnclaveId, requestedUserServiceGuids, requestedShouldFollowLogs)
 		require.NoError(t, err)
 		require.NotNil(t, cancelStreamUserServiceLogsFunc)
 		require.NotNil(t, userServiceLogsByGuidChan)
 
 		receivedLogLines := []string{}
+		receivedNotFoundUserServiceGuids := map[services.ServiceGUID]bool{}
 
 		var testEvaluationErr error
 
-		shouldReceiveStream := true
-		for shouldReceiveStream {
+		shouldReceiveUserServiceLogs := true
+		shouldReceiveNonFoundUserServiceGuids := true
+
+		for shouldReceiveUserServiceLogs || shouldReceiveNonFoundUserServiceGuids {
 			select {
 			case <-time.Tick(testTimeOut):
 				testEvaluationErr = stacktrace.NewError("Receiving stream logs in the test has reached the '%v' time out", testTimeOut)
-				shouldReceiveStream = false
+				shouldReceiveUserServiceLogs = false
+				shouldReceiveNonFoundUserServiceGuids = false
 				break
 			case userServiceLogsByGuid, isChanOpen := <-userServiceLogsByGuidChan:
 				if !isChanOpen {
-					shouldReceiveStream = false
+					shouldReceiveUserServiceLogs = false
+					shouldReceiveNonFoundUserServiceGuids = false
 					break
 				}
 
@@ -112,18 +154,30 @@ func TestStreamLogs(t *testing.T) {
 					receivedLogLines = append(receivedLogLines, serviceLog.GetContent())
 				}
 
-				if len(receivedLogLines) == expectedAmountOfLogLines {
-					shouldReceiveStream = false
+				if len(receivedLogLines) == len(expectedLogLines) {
+					shouldReceiveUserServiceLogs = false
 					break
 				}
+			case notFoundGuids, isChanOpen := <-notFoundUserServiceGuidsChan:
+				if !isChanOpen {
+					shouldReceiveUserServiceLogs = false
+					shouldReceiveNonFoundUserServiceGuids = false
+					break
+				}
+
+				receivedNotFoundUserServiceGuids = notFoundGuids
+				shouldReceiveNonFoundUserServiceGuids = false
+				break
 			}
 		}
 
 		require.NoError(t, testEvaluationErr)
 		require.Equal(t, expectedLogLines, receivedLogLines)
+		require.Equal(t, expectedNonExistenceUserServiceGuids, receivedNotFoundUserServiceGuids)
 		cancelStreamUserServiceLogsFunc()
 	}
 }
+
 
 // ====================================================================================================
 //                                       Private helper functions
@@ -131,7 +185,7 @@ func TestStreamLogs(t *testing.T) {
 func getExampleServiceConfig() *services.ContainerConfig {
 
 	entrypointArgs := []string{"/bin/sh", "-c"}
-	cmdArgs := []string{"for i in kurtosis test running successfully; do echo \"$i\"; if [ \"$i\" == \"successfully\" ]; then sleep 300; fi; done;"}
+	cmdArgs := []string{"for i in kurtosis test running successfully; do echo \"$i\"; done;"}
 
 	containerConfig := services.NewContainerConfigBuilder(
 		dockerGettingStartedImage,
@@ -139,8 +193,6 @@ func getExampleServiceConfig() *services.ContainerConfig {
 		entrypointArgs,
 	).WithCmdOverride(
 		cmdArgs,
-	).WithUsedPorts(map[string]*services.PortSpec{
-		exampleServicePortId: exampleServicePrivatePortSpec,
-	}).Build()
+	).Build()
 	return containerConfig
 }